+++ crd-generator/api/src/test/java/io/fabric8/crd/example/multiple/v1/Multiple.java
@@ -13,8 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.model.annotation;
+package io.fabric8.crd.example.multiple.v1;
 
-public @interface PackageSuffix {
-  String value();
+import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.model.annotation.Group;
+import io.fabric8.kubernetes.model.annotation.Version;
+
+@Group("sample.fabric8.io")
+@Version("v1")
+public class Multiple extends CustomResource<MultipleSpec, Void> {
 }