+++ extensions/knative/generator/cmd/generate/generate.go
@@ -136,7 +136,7 @@ func main() {
 		reflect.TypeOf(runtime.RawExtension{}): "java.util.Map<String, Object>",
 	}
 
-	json := schemagen.GenerateSchema("http://fabric8.io/knative/KnativeSchema#", crdLists, providedPackages, manualTypeMap, packageMapping, mappingSchema, providedTypes, constraints)
+	json := schemagen.GenerateSchema("http://fabric8.io/knative/KnativeSchema#", crdLists, providedPackages, manualTypeMap, packageMapping, mappingSchema, providedTypes, constraints, "io.fabric8")
 
 	fmt.Println(json)
 }