+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/BaseKubernetesClient.java
@@ -94,8 +94,7 @@ import io.fabric8.kubernetes.client.dsl.Waitable;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.base.ResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.internal.ClusterOperationsImpl;
-import io.fabric8.kubernetes.client.dsl.internal.CreateOnlyResourceOperationsImpl;
-import io.fabric8.kubernetes.client.dsl.internal.ResourceOperationsImpl;
+import io.fabric8.kubernetes.client.dsl.internal.HasMetadataOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.apps.v1.DeploymentOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.apps.v1.ReplicaSetOperationsImpl;
 import io.fabric8.kubernetes.client.dsl.internal.apps.v1.StatefulSetOperationsImpl;
@@ -394,7 +393,7 @@ public abstract class BaseKubernetesClient<C extends Client> extends BaseClient 
    */
   @Override
   public InOutCreateable<TokenReview, TokenReview> tokenReviews() {
-    return new CreateOnlyResourceOperationsImpl<>(httpClient, getConfiguration(), "authentication.k8s.io", "v1", HasMetadata.getPlural(TokenReview.class), TokenReview.class);
+    return Handlers.getNonListingOperation(TokenReview.class, httpClient, getConfiguration());
   }
 
   /**
@@ -415,7 +414,7 @@ public abstract class BaseKubernetesClient<C extends Client> extends BaseClient 
   }
   
   @Override
-  public <T extends HasMetadata, L extends KubernetesResourceList<T>> ResourceOperationsImpl<T, L> resources(
+  public <T extends HasMetadata, L extends KubernetesResourceList<T>> HasMetadataOperationsImpl<T, L> resources(
       Class<T> resourceType, Class<L> listClass) {
     return customResources(ResourceDefinitionContext.fromResourceType(resourceType), resourceType, listClass);
   }
@@ -424,8 +423,8 @@ public abstract class BaseKubernetesClient<C extends Client> extends BaseClient 
    * {@inheritDoc}
    */
   @Override
-  public <T extends HasMetadata, L extends KubernetesResourceList<T>> ResourceOperationsImpl<T, L> customResources(ResourceDefinitionContext rdContext, Class<T> resourceType, Class<L> listClass) {
-    return new ResourceOperationsImpl<>(httpClient, getConfiguration(), rdContext, resourceType, listClass);
+  public <T extends HasMetadata, L extends KubernetesResourceList<T>> HasMetadataOperationsImpl<T, L> customResources(ResourceDefinitionContext rdContext, Class<T> resourceType, Class<L> listClass) {
+    return new HasMetadataOperationsImpl<>(httpClient, getConfiguration(), rdContext, resourceType, listClass);
   }
 
   /**